+++ weatherTests/weatherTests.swift
@@ -9,7 +9,7 @@ import XCTest
 @testable import weather
 
 struct ListMock: DateContaining, Equatable {
-    let dt: Double
+    let dt: Int
 }
 
 
@@ -19,17 +19,17 @@ class weatherTests: XCTestCase {
 
         let list: [ListMock] = [
 
-            ListMock(dt: 1643835600.0),
-            ListMock(dt: 1643857200.0),
-            ListMock(dt: 1643922000.0)
+            ListMock(dt: 1643835600),
+            ListMock(dt: 1643857200),
+            ListMock(dt: 1643922000)
         ]
 
         //not optimal because function lives everywhere but okay for Linas knowledge right now
         let filteredList = filterNoon(unfilteredList: list)
         let expectedList: [ListMock] = [
 
-            ListMock(dt: 1643835600.0),
-            ListMock(dt: 1643922000.0)
+            ListMock(dt: 1643835600),
+            ListMock(dt: 1643922000)
         ]
 
         XCTAssertEqual(filteredList, expectedList)
